add config, template, repo, fix fabfile

+++ fabfile.py
@@ -7,9 +7,10 @@ Usage:
 """
 from fabric.api import *
 
-env.hosts = ['web_server', 'localhost']
+env.hosts = ['localhost']
 env.user = 'root'
-env.port = 7878
+env.password = 'ABC@123'
+env.port = 22
 #env.use_ssh_config = True
 env.key_filename = ['/home/keys/prikey.pem']
 
@@ -42,15 +43,18 @@ def setup_nginx():
     run('service nginx restart')
 
 def setup_php():
-    run('rpm -Uvh http://rpms.famillecollet.com/enterprise/remi-release-7.rpm')
-    run('yum --enablerepo=remi,remi-php56 install -y php-opcache php php-fpm php-mysql php-intl php-mcrypt php-gd')
+    #run('rpm -Uvh http://rpms.famillecollet.com/enterprise/remi-release-7.rpm')
+    upload_remi_repo = put('yum.repos.d/remi.repo', '/etc/yum.repos.d/remi.repo')
 
-    # Config pool www in php-fpm
-    run('rm -rf /etc/php-fpm.d/www.conf')
-    upload_www_conf = put('config/www.conf', '/etc/php-fpm.d/www.conf')
-    if upload_www_conf.succeeded:
-	print "Upload www.conf successfully."
-	run('service php-fpm restart')
+    if upload_remi_repo.succeeded:
+    	run('yum install -y php-opcache php php-fpm php-mysql php-intl php-mcrypt php-gd')
+
+    	# Config pool www in php-fpm
+    	run('rm -rf /etc/php-fpm.d/www.conf')
+    	upload_www_conf = put('config/www.conf', '/etc/php-fpm.d/www.conf')
+    	if upload_www_conf.succeeded:
+		print "Upload www.conf successfully."
+		run('service php-fpm restart')
 
 
 def setup_mariadb(galera=False, my_level='medium'):
@@ -60,10 +64,10 @@ def setup_mariadb(galera=False, my_level='medium'):
 
     if upload_mariadb_repo.succeeded:
         print "Upload successfully ..."
-        if galera == False:
-            run('yum install -y MariaDB-server MariaDB-client')
+        if galera:
+            run('yum install -y MariaDB-server MariaDB-client galera')
         else:
-            run('yum install -y MariaDB-Galera-server MariaDB-client galera')
+            run('yum install -y MariaDB-Galera-server MariaDB-client')
 
 	run('rm -rf /etc/my.cnf')
         if my_level == 'medium':
@@ -77,12 +81,14 @@ def setup_mariadb(galera=False, my_level='medium'):
 
 def deploy(nginx=True, php=True, mariadb=False, galera=False, my_level='medium'):
 
-    if nginx == True:
+    if nginx:
     	setup_nginx()
+	print "test"
 
-    if php == True:
+    if php:
 	setup_php()
 
-    if mariadb == True:
+    if mariadb:
 	setup_mariadb(galera, my_level)
 
+    print "TEST"
